datapath: fix lintall issues after ids() refactor
- Remove trailing blank line inside test module in pool.rs
- Remove unnecessary u64->u64 cast in controller/service.rs (id was
  already u64 after the pool API change)

+++ data-plane/core/controller/src/service.rs
@@ -703,7 +703,7 @@ impl ControllerService {
             .connection_table()
             .for_each(|id, conn| {
                 if conn.link_id().as_deref() == Some(link_id) && resolved.is_none() {
-                    resolved = Some(id as u64);
+                    resolved = Some(id);
                 }
             });
